pyproject.toml
[tool.poetry]
name = "graphrag"
version = "0.1.0"
description = "GraphRAG PoC with LangChain, LlamaIndex, Neo4j, and FAISS"
authors = ["Brian Brewer <bbrewer@infolibcorp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
# LangChain ecosystem
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-ollama = "^0.3.8"
# LlamaIndex
llama-index = "^0.14.0"
# Neo4j graph database
neo4j = "^6.0.0"
# Vector search
faiss-cpu = "^1.13.0"
sentence-transformers = "^2.7.0"
# Data & utilities
pandas = "^2.2.2"
numpy = "^1.24.0"
faker = "^19.0.0"
scikit-learn = "^1.5.0"
# App framework
streamlit = "^1.31.0"
# Helpers
python-dotenv = "^1.0.0"
pyyaml = "^6.0.0"
tiktoken = "^0.7.0"
networkx = "^3.0"
unstructured = "^0.14.0"
beautifulsoup4 = "^4.12.2"
# LLM clients
openai = "^1.0.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ipykernel = "^6.30.0"
jupyterlab = "^4.0.0"
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.6.0"
mypy = "^1.13.0"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
